_archive/diachronic-valency-corpus/requirements.txt
psycopg2-binary
requests
stanza
schedule
fastapi
uvicorn
asyncpg
aiofiles
pandas
beautifulsoup4
lxml
nltk
spacy
transformers
torch
numpy
python-multipart
